Lexical variables work, ast exprs now have TYPE
Before, forms waited to evaluate their arguments until TO-INSTRUCTIONS,
but now they evaluate them immediately, store the AST-EXPRs, and then
call TO-INSTRUCTIONS on them later. This allows each AST-EXPR to know
what type it is before generating instructions, which is pretty
necessary to do anything.
Forms now have to specify their type as well, instead of just putting it
into the INSTR-RESULT return value of TO-INSTRUCTIONS.